# router, backend: replace time.Time with mono-time to optimize duration calculation