<!--

Thank you for contributing to TiProxy!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and 
linking the relevant issues via the "close: #xxx" or "ref: #xxx".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: ref #420 

Problem Summary:
We're adding more and more metrics and the overhead of time.Now() is getting more.
time.Now() involves wall time and nanotime, but we only need nanotime to calculate duration.

What is changed and how it works:
Replace time.Time with monotime.Time.

### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [ ] Integration test
- [ ] Manual test (add detailed scripts or steps below)
- [ ] No code

Benchmark results:

```
BenchmarkGoNow
BenchmarkGoNow-8       	13206856	        82.03 ns/op
BenchmarkMonoNow
BenchmarkMonoNow-8     	33960902	        34.73 ns/op
BenchmarkGoSince
BenchmarkGoSince-8     	 9926468	       123.3 ns/op
BenchmarkMonoSince
BenchmarkMonoSince-8   	15168806	        66.29 ns/op
```

Notable changes

- [ ] Has configuration change
- [ ] Has HTTP API interfaces change
- [ ] Has tiproxyctl change
- [ ] Other user behavior changes

### Release note

<!-- compatibility change, improvement, bugfix, and new feature need a release note -->

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

```release-note
Replace time.Time with mono-time to optimize duration calculation
```